requirements.txt
roboflow==1.2.0
opencv-python==4.8.0.76
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0
requests==2.31.0
tqdm==4.65.0
seaborn==0.12.2
ultralytics>=8.0.196
torch>=2.0.0
torchvision>=0.15.0
# YOLOv11 support
